requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Migrate==1.3.0
Flask-Script==2.0.5
gunicorn==19.2.1
Markdown==2.5.2
psycopg2==2.6
requests==2.5.3
Sphinx==1.2.3
